pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tgvectordb"
version = "0.1.0"
description = "Free, unlimited vector database backed by Telegram"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "TgVectorDB Contributors"}
]
keywords = ["vector-database", "telegram", "embeddings", "rag", "semantic-search"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "telethon>=1.36.0",
    "cryptg",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
pdf = ["pdfplumber>=0.10.0"]
docx = ["python-docx>=1.0.0"]
all = ["pdfplumber>=0.10.0", "python-docx>=1.0.0"]
dev = ["pytest", "pytest-asyncio"]

[project.urls]
Homepage = "https://github.com/yourname/tgvectordb"
Issues = "https://github.com/yourname/tgvectordb/issues"
